Update package to last version of rcomponent
