+++ .gitignore
@@ -1,1 +1,8 @@
+# Ignore pyc files
 *.pyc
+
+# Ignore vscode folder
+.vscode/
+
+# Ignore temporal files from .deb generation
+*.ddeb
