+++ CHANGELOG.rst
@@ -1,0 +1,12 @@
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package roboteq_bms
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Forthcoming
+-----------
+
+0.1.0 (2020-XX-XX)
+------------------
+* Work in progress...
+*
+* Contributors: Jose Antonio Mendez
